don't copy templates folder if it doesn't exist

+++ devtools/dev_env.py
@@ -69,8 +69,9 @@ def Build(args):
     shutil.copytree(BACKEND_DIR, BUILD_DIR, dirs_exist_ok=True,
                     ignore=shutil.ignore_patterns('*.pyc', '__pycache__'))
     shutil.copy2(REQUIREMENTS_PATH, os.path.join(BUILD_DIR, 'requirements.txt'))
-    shutil.copytree(os.path.join(ROOT_DIR, 'templates'),
-                    os.path.join(BUILD_DIR, 'templates'))
+    if os.path.exists(os.path.join(ROOT_DIR, 'templates')):
+      shutil.copytree(os.path.join(ROOT_DIR, 'templates'),
+                      os.path.join(BUILD_DIR, 'templates'))
     shutil.copytree(FRONTEND_DIST_DIR, BUILD_STATIC_DIR)
